pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-tracker"
version = "0.1.0"
description = "AI-powered personal finance tracker CLI"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Your Name"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: Financial",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
]

[project.scripts]
finance = "src.cli.main:cli"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.ruff]
line-length = 120
target-version = "py39"
